Began work on schedule component.
This includes just making sure the component shows the schedule saved
from the database along with general formatting of the schedule display.
This includes the names of the days along with different times for each
day. All of this is stored in a javascript object that maps days to 2d
arrays of booleans that indicate which hour time block the event is scheduled.