Fixed issues with schedule state not updating and clearing a day's schedule.
I needed to make a copy for every time I change a value within the array which
is 96 x 7. There may not be a way of fixing this but we'll have to see. Also the
clear schedule for the day works now. This is also 96 x 7 plus another state update.
Very slow.