Finished Project.
With how the project is designed I could go about remodeling it and
making a scheduling app that would be for multiple users which would be more fitting.
Only problem would be that this would be that it requires a ground up rebuild of the
data stored and new components. The other would be a more realistic scheduler that lets
you pick the times from a drop down menu and give the user the ability to title events
and give descriptions. This would be a better design rather than the click and drag
scheduling app I have made. Food for thought.